requirements.txt
sanic
sanic-ext
python-dotenv
sqlalchemy[asyncio]
asyncpg
pydantic>=1.9.0
python-jose[cryptography]
setuptools<81.0.0
pyjwt>=2.0.0
werkzeug>=2.0.0
setuptools<81
setuptools<81
aiofiles==24.1.0
annotated-types==0.7.0
asyncpg==0.30.0
cffi==1.17.1
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
greenlet==3.2.2
html5tagger==1.3.0
httptools==0.6.4
idna==3.10
MarkupSafe==3.0.2
multidict==6.4.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
sanic==25.3.0
sanic-ext==24.12.0
sanic-routing==23.12.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
tracerite==1.1.1
typing-inspection==0.4.1
typing_extensions==4.13.2
ujson==5.10.0
uvloop==0.21.0
websockets==15.0.1
Werkzeug==3.1.3
